Move interfaces to types.go (best practice)
